refactor(clipboard): replace direct navigator.clipboard calls with resilient clipboard utility (Fixes #611)
