+++ app/[locale]/insights/page.js
@@ -18,6 +18,7 @@ import { useTranslations } from 'next-intl'
 import SymptomPhaseInsights from '@/components/dashboard/SymptomPhaseInsights'
 import { formatDateForCSV } from '@/lib/utils'
 import { normaliseRiskResult } from '@/lib/pcod-risk-result'
+import { copyToClipboard } from '@/lib/clipboard'
 import SectionCard, { IconBadge } from '@/components/ui/SectionCard'
 
 const WeightTrendChart = dynamic(() => import('@/components/dashboard/WeightTrendChart'), {
@@ -265,31 +266,10 @@ export default function InsightsPage() {
 - Recent PCOD Risk: ${riskLevelWord}
 - Logged Symptoms (Last 30 Days): ${recentSymptomsText}`
 
-    try {
-      if (navigator.clipboard && typeof navigator.clipboard.writeText === 'function') {
-        await navigator.clipboard.writeText(summaryText)
-        setCopied(true)
-        setTimeout(() => setCopied(false), 2000)
-      } else {
-        const textArea = document.createElement('textarea')
-        textArea.value = summaryText
-        textArea.style.position = 'fixed'
-        textArea.style.left = '-999999px'
-        textArea.style.top = '-999999px'
-        document.body.appendChild(textArea)
-        textArea.focus()
-        textArea.select()
-        const successful = document.execCommand('copy')
-        document.body.removeChild(textArea)
-        if (successful) {
-          setCopied(true)
-          setTimeout(() => setCopied(false), 2000)
-        } else {
-          throw new Error('Fallback copy failed')
-        }
-      }
-    } catch (err) {
-      console.error('Could not copy summary', err)
+    const ok = await copyToClipboard(summaryText)
+    if (ok) {
+      setCopied(true)
+      setTimeout(() => setCopied(false), 2000)
     }
   }
